Move extern_crate_name to Resolve
